Finish concept

+++ app/src/main/java/com/example/teacherassistant/Model/Repositories/CourseRepository.kt
@@ -1,5 +1,6 @@
 package com.example.teacherassistant.Model.Repositories
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import com.example.teacherassistant.Model.Course
 import com.example.teacherassistant.Model.Dao.CourseDao
@@ -10,12 +11,19 @@ class CourseRepository(private val courseDao: CourseDao) {
 
     suspend fun addCourse(course: Course){
         courseDao.insertCourse(course)
+        Log.v("course", "added")
     }
-    suspend fun removeCourse(course: Course){
-        courseDao.deleteCourse(course)
+    suspend fun removeCourse(course: Course?){
+        if (course != null){
+            courseDao.deleteCourse(course)
+        }
+        Log.v("course", "removed")
     }
-    suspend fun updateCourse(course: Course){
-        courseDao.updateCourse(course)
+    suspend fun updateCourse(course: Course?){
+        if(course != null){
+            courseDao.updateCourse(course)
+        }
+        Log.v("course", "updated")
     }
-    fun getStudentById(id: Int): Course = courseDao.getCourseById(id)
+    fun getCourseById(idc: Int): Course = courseDao.getCourseById(idc)
 }
